Fix rentals code
